Add a Tor Transport
Add a Tor transport based on the Tokio TCP Transport. This Transport
allows users of libp2p to configure an onion service via Tor.

For the listener side of the application, using this transport, one
can pass an onion address to the application. libp2p will pass this
address down to the transport via the call to `listen_on`. In order to
create a listening socket to redirect the onion service to we must be
able to map from the onion `Multiaddr` to the port number to use. In
order to do this we add a `HashMap` to the `TorTokioTcpConfig`, the
user must set up this map when building the transport.

For the dialer side things are considerably more simple. When the
`dial` method is called we set up an IPv4 TCP connection to the
locally running Tor instance's SOCKS5 proxy. We then simply return the
`TcpStream` that was created.

This works with upgrades and is tested with multiplexing via yamux and
authentication via secio.

Introduces a lot of duplicate code between the TCP transport and this one.